requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-action-serializer==1.0.5
drf-yasg==1.20.0
geographiclib==1.50
geopy==2.1.0
gunicorn==20.1.0
haversine==2.3.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
phonenumbers==8.12.21
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==2.0.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
